reorganize directory structure and add more help to README
