+++ README.md
@@ -6,6 +6,13 @@ Initialized with https://github.com/Microsoft/TypeScript-React-Starter. Check th
 Info about React: https://reactjs.org/tutorial/tutorial.html
 Info about TypeScript: https://www.typescriptlang.org/docs/handbook/typescript-in-5-minutes.html
 
+# Organization
+`momentum/index.tsx` is the main file
+`momentum/components` contains components which can be include in `index.tsx`
+`momentum/data` contains data that components and `index.tsx` can use
+
+`momentum/components/App` contains an example component
+
 # How to get setup (assuming macOS).
 1. Install HomeBrew
 2. Install Node: `brew install node`
@@ -17,6 +24,6 @@ The project was tested using VSCode . YMMV with other tools.
 
 # Running
 From `/momentum`:
-- `npm start` starts the dev server (use this for development)
+- `npm start` starts the dev server (use this for development). the site will update automatically when you edit files
 - `npm run build` compiles an optimized version of project (use this for deployment)
 - `npm test` runs tests
