+++ momentum/src/components/App/App.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import './App.css';
 
-import logo from './logo.svg';
+import logo from '../../data/logo.svg';
 
 class App extends React.Component {
   public render() {
@@ -12,7 +12,7 @@ class App extends React.Component {
           <h1 className="App-title">Welcome to React</h1>
         </header>
         <p className="App-intro">
-          To get started, edit <code>src/App.tsx</code> and save to reload.
+          To get started, edit <code>src/components/App/App.tsx</code> and save to reload.
         </p>
       </div>
     );
