+++ momentum/src/index.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
-import App from './App';
+import App from './components/App/App';
 import './index.css';
 import registerServiceWorker from './registerServiceWorker';
